KEYCLOAK-18489 Extract methods from models with multi-valued attributes into interface
