+++ model/infinispan/src/main/java/org/keycloak/models/cache/infinispan/RoleAdapter.java
@@ -175,12 +175,6 @@ public class RoleAdapter implements RoleModel {
     }
 
     @Override
-    public void setSingleAttribute(String name, String value) {
-        getDelegateForUpdate();
-        updated.setSingleAttribute(name, value);
-    }
-
-    @Override
     public void setAttribute(String name, List<String> values) {
         getDelegateForUpdate();
         updated.setAttribute(name, values);
