+++ server-spi/src/main/java/org/keycloak/models/RoleModel.java
@@ -17,6 +17,7 @@
 
 package org.keycloak.models;
 
+import java.util.Collections;
 import org.keycloak.storage.SearchableModelField;
 import java.util.List;
 import java.util.Map;
@@ -28,7 +29,7 @@ import java.util.stream.Stream;
  * @author <a href="mailto:bill@burkecentral.com">Bill Burke</a>
  * @version $Revision: 1 $
  */
-public interface RoleModel {
+public interface RoleModel extends ModelWithAttributes {
 
     public static class SearchableFields {
         public static final SearchableModelField<RoleModel> ID                  = new SearchableModelField<>("id", String.class);
@@ -79,16 +80,6 @@ public interface RoleModel {
 
     boolean hasRole(RoleModel role);
 
-    void setSingleAttribute(String name, String value);
-
-    void setAttribute(String name, List<String> values);
-
-    void removeAttribute(String name);
-
-    default String getFirstAttribute(String name) {
-        return getAttributeStream(name).findFirst().orElse(null);
-    }
-
     /**
      * @deprecated Use {@link #getAttributeStream(String) getAttributeStream} instead.
      */
@@ -96,13 +87,4 @@ public interface RoleModel {
     default List<String> getAttribute(String name) {
         return getAttributeStream(name).collect(Collectors.toList());
     }
-
-    /**
-     * Returns all role's attributes that match the given name as a stream.
-     * @param name {@code String} Name of an attribute to be used as a filter.
-     * @return Stream of {@code String}. Never returns {@code null}.
-     */
-    Stream<String> getAttributeStream(String name);
-
-    Map<String, List<String>> getAttributes();
 }
