+++ server-spi/src/main/java/org/keycloak/models/UserModel.java
@@ -23,7 +23,6 @@ import org.keycloak.storage.SearchableModelField;
 import java.util.Comparator;
 import java.util.LinkedHashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -32,7 +31,7 @@ import java.util.stream.Stream;
  * @author <a href="mailto:bill@burkecentral.com">Bill Burke</a>
  * @version $Revision: 1 $
  */
-public interface UserModel extends RoleMapperModel {
+public interface UserModel extends RoleMapperModel, ModelWithAttributes {
     String USERNAME = "username";
     String FIRST_NAME = "firstName";
     String LAST_NAME = "lastName";
@@ -126,24 +125,6 @@ public interface UserModel extends RoleMapperModel {
     void setEnabled(boolean enabled);
 
     /**
-     * Set single value of specified attribute. Remove all other existing values of this attribute
-     *
-     * @param name
-     * @param value
-     */
-    void setSingleAttribute(String name, String value);
-
-    void setAttribute(String name, List<String> values);
-
-    void removeAttribute(String name);
-
-    /**
-     * @param name
-     * @return null if there is not any value of specified attribute or first value otherwise. Don't throw exception if there are more values of the attribute
-     */
-    String getFirstAttribute(String name);
-
-    /**
      * @param name
      * @return list of all attribute values or empty list if there are not any values. Never return null
      * @deprecated Use {@link #getAttributeStream(String) getAttributeStream} instead.
@@ -157,12 +138,11 @@ public interface UserModel extends RoleMapperModel {
      * @param name the name of the attribute.
      * @return a non-null {@link Stream} of attribute values.
      */
+    @Override
     default Stream<String> getAttributeStream(final String name) {
         List<String> value = this.getAttribute(name);
         return value != null ? value.stream() : Stream.empty();
     }
-
-    Map<String, List<String>> getAttributes();
 
     /**
      * @deprecated Use {@link #getRequiredActionsStream() getRequiredActionsStream} instead.
